Move config out to another file
